move commercial knex files to shared
